feat: add tabs for processing view
