+++ crimson/src/crimson/etl_modules/ImportSingleCsvWithProcessing/components/ErrorPreviewComponent.vue
@@ -1,0 +1,13 @@
+<template>
+    <div>
+        <h1>Error Checking View</h1>
+    </div>
+</template>
+
+<script setup>
+
+</script>
+
+<style scoped>
+
+</style>
